+++ inc/detection.h
@@ -6,7 +6,7 @@
 
 #include <opencv2/opencv.hpp>
 
-typedef std::vector<std::tuple<std::vector<std::tuple<double, double>>, std::vector<cv::Mat>>> MSOPDescriptor;
+typedef std::vector<std::tuple<std::vector<std::tuple<float, float>>, std::vector<cv::Mat>>> MSOPDescriptor;
 
 MSOPDescriptor get_MSOP_features(const cv::Mat&);
